refactor: Consolidate BarChart
