+++ src/charts/index.ts
@@ -1,0 +1,1 @@
+export * as BarChart from "./BarChart";
